+++ internal/store/k8s.go
@@ -1,4 +1,5 @@
 // Copyright 2018 Sorint.lab
+// Copyright 2026 WoozyMasta
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -49,14 +50,14 @@ const (
 )
 
 type KubeStore struct {
-	client       *kubernetes.Clientset
+	client       kubernetes.Interface
 	podName      string
 	namespace    string
 	clusterName  string
 	resourceName string
 }
 
-func NewKubeStore(kubecli *kubernetes.Clientset, podName, namespace, clusterName string) (*KubeStore, error) {
+func NewKubeStore(kubecli kubernetes.Interface, podName, namespace, clusterName string) (*KubeStore, error) {
 	return &KubeStore{
 		client:       kubecli,
 		podName:      podName,
@@ -74,10 +75,10 @@ func (s *KubeStore) labelSelector(componentLabel ComponentLabelValue) labels.Sel
 	return labels.SelectorFromSet(selector)
 }
 
-func (s *KubeStore) patchKubeStatusAnnotation(annotationData []byte) error {
+func (s *KubeStore) patchKubeStatusAnnotation(ctx context.Context, annotationData []byte) error {
 	podsClient := s.client.CoreV1().Pods(s.namespace)
 	retryErr := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		pod, err := podsClient.Get(s.podName, metav1.GetOptions{})
+		pod, err := podsClient.Get(ctx, s.podName, metav1.GetOptions{})
 		if err != nil {
 			return fmt.Errorf("failed to get latest version of pod: %v", err)
 		}
@@ -102,7 +103,7 @@ func (s *KubeStore) patchKubeStatusAnnotation(annotationData []byte) error {
 			return fmt.Errorf("failed to create pod merge patch: %v", err)
 		}
 
-		_, err = podsClient.Patch(s.podName, types.MergePatchType, patchBytes)
+		_, err = podsClient.Patch(ctx, s.podName, types.MergePatchType, patchBytes, metav1.PatchOptions{})
 		return err
 	})
 	if retryErr != nil {
@@ -119,7 +120,7 @@ func (s *KubeStore) AtomicPutClusterData(ctx context.Context, cd *cluster.Cluste
 	epsClient := s.client.CoreV1().ConfigMaps(s.namespace)
 
 	retryErr := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		result, err := epsClient.Get(s.resourceName, metav1.GetOptions{})
+		result, err := epsClient.Get(ctx, s.resourceName, metav1.GetOptions{})
 		if err != nil && !apierrors.IsNotFound(err) {
 			return fmt.Errorf("failed to get latest version of configmap: %v", err)
 		}
@@ -158,7 +159,7 @@ func (s *KubeStore) AtomicPutClusterData(ctx context.Context, cd *cluster.Cluste
 				result.Annotations = map[string]string{}
 			}
 			result.Annotations[util.KubeClusterDataAnnotation] = string(cdj)
-			_, err = epsClient.Update(result)
+			_, err = epsClient.Update(ctx, result, metav1.UpdateOptions{})
 			return err
 		} else {
 			// configmap does not exists
@@ -168,12 +169,12 @@ func (s *KubeStore) AtomicPutClusterData(ctx context.Context, cd *cluster.Cluste
 				return ErrKeyModified
 			}
 			annotations := map[string]string{util.KubeClusterDataAnnotation: string(cdj)}
-			_, err = epsClient.Create(&v1.ConfigMap{
+			_, err = epsClient.Create(ctx, &v1.ConfigMap{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:        s.resourceName,
 					Annotations: annotations,
 				},
-			})
+			}, metav1.CreateOptions{})
 			return err
 		}
 	})
@@ -191,7 +192,7 @@ func (s *KubeStore) PutClusterData(ctx context.Context, cd *cluster.ClusterData)
 	epsClient := s.client.CoreV1().ConfigMaps(s.namespace)
 
 	retryErr := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		result, err := epsClient.Get(s.resourceName, metav1.GetOptions{})
+		result, err := epsClient.Get(ctx, s.resourceName, metav1.GetOptions{})
 		if err != nil && !apierrors.IsNotFound(err) {
 			return fmt.Errorf("failed to get latest version of configmap: %v", err)
 		}
@@ -201,17 +202,17 @@ func (s *KubeStore) PutClusterData(ctx context.Context, cd *cluster.ClusterData)
 				result.Annotations = map[string]string{}
 			}
 			result.Annotations[util.KubeClusterDataAnnotation] = string(cdj)
-			_, err = epsClient.Update(result)
+			_, err = epsClient.Update(ctx, result, metav1.UpdateOptions{})
 			return err
 		} else {
 			// configmap does not exists
 			annotations := map[string]string{util.KubeClusterDataAnnotation: string(cdj)}
-			_, err = epsClient.Create(&v1.ConfigMap{
+			_, err = epsClient.Create(ctx, &v1.ConfigMap{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:        s.resourceName,
 					Annotations: annotations,
 				},
-			})
+			}, metav1.CreateOptions{})
 			return err
 		}
 	})
@@ -223,7 +224,7 @@ func (s *KubeStore) PutClusterData(ctx context.Context, cd *cluster.ClusterData)
 
 func (s *KubeStore) GetClusterData(ctx context.Context) (*cluster.ClusterData, *KVPair, error) {
 	epsClient := s.client.CoreV1().ConfigMaps(s.namespace)
-	result, err := epsClient.Get(s.resourceName, metav1.GetOptions{})
+	result, err := epsClient.Get(ctx, s.resourceName, metav1.GetOptions{})
 	if err != nil {
 		if apierrors.IsNotFound(err) {
 			return nil, nil, nil
@@ -249,7 +250,7 @@ func (s *KubeStore) SetKeeperInfo(ctx context.Context, id string, ms *cluster.Ke
 	if err != nil {
 		return err
 	}
-	return s.patchKubeStatusAnnotation(msj)
+	return s.patchKubeStatusAnnotation(ctx, msj)
 }
 
 func (s *KubeStore) GetKeepersInfo(ctx context.Context) (cluster.KeepersInfo, error) {
@@ -260,7 +261,7 @@ func (s *KubeStore) GetKeepersInfo(ctx context.Context) (cluster.KeepersInfo, er
 	listOpts := metav1.ListOptions{
 		LabelSelector: s.labelSelector(KeeperLabelValue).String(),
 	}
-	result, err := podsClient.List(listOpts)
+	result, err := podsClient.List(ctx, listOpts)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get latest version of pod: %v", err)
 	}
@@ -284,7 +285,7 @@ func (s *KubeStore) SetSentinelInfo(ctx context.Context, si *cluster.SentinelInf
 	if err != nil {
 		return err
 	}
-	return s.patchKubeStatusAnnotation(sij)
+	return s.patchKubeStatusAnnotation(ctx, sij)
 }
 
 func (s *KubeStore) GetSentinelsInfo(ctx context.Context) (cluster.SentinelsInfo, error) {
@@ -295,7 +296,7 @@ func (s *KubeStore) GetSentinelsInfo(ctx context.Context) (cluster.SentinelsInfo
 	listOpts := metav1.ListOptions{
 		LabelSelector: s.labelSelector(SentinelLabelValue).String(),
 	}
-	result, err := podsClient.List(listOpts)
+	result, err := podsClient.List(ctx, listOpts)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get latest version of pod: %v", err)
 	}
@@ -319,7 +320,7 @@ func (s *KubeStore) SetProxyInfo(ctx context.Context, pi *cluster.ProxyInfo, ttl
 	if err != nil {
 		return err
 	}
-	return s.patchKubeStatusAnnotation(pij)
+	return s.patchKubeStatusAnnotation(ctx, pij)
 }
 
 func (s *KubeStore) GetProxiesInfo(ctx context.Context) (cluster.ProxiesInfo, error) {
@@ -330,7 +331,7 @@ func (s *KubeStore) GetProxiesInfo(ctx context.Context) (cluster.ProxiesInfo, er
 	listOpts := metav1.ListOptions{
 		LabelSelector: s.labelSelector(ProxyLabelValue).String(),
 	}
-	result, err := podsClient.List(listOpts)
+	result, err := podsClient.List(ctx, listOpts)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get latest version of pod: %v", err)
 	}
@@ -350,7 +351,7 @@ func (s *KubeStore) GetProxiesInfo(ctx context.Context) (cluster.ProxiesInfo, er
 }
 
 type KubeElection struct {
-	client       *kubernetes.Clientset
+	client       kubernetes.Interface
 	podName      string
 	namespace    string
 	resourceName string
@@ -366,14 +367,14 @@ type KubeElection struct {
 	rl resourcelock.Interface
 }
 
-func NewKubeElection(kubecli *kubernetes.Clientset, podName, namespace, clusterName, candidateUID string) (*KubeElection, error) {
+func NewKubeElection(kubecli kubernetes.Interface, podName, namespace, clusterName, candidateUID string) (*KubeElection, error) {
 	resourceName := fmt.Sprintf("%s-%s", util.KubeResourcePrefix, clusterName)
 
-	rl, err := resourcelock.New(resourcelock.ConfigMapsResourceLock,
+	rl, err := resourcelock.New(resourcelock.LeasesResourceLock,
 		namespace,
 		resourceName,
 		kubecli.CoreV1(),
-		nil,
+		kubecli.CoordinationV1(),
 		resourcelock.ResourceLockConfig{
 			Identity:      candidateUID,
 			EventRecorder: createRecorder(kubecli, "stolon-sentinel", namespace),
@@ -415,7 +416,7 @@ func (e *KubeElection) Stop() {
 }
 
 func (e *KubeElection) Leader() (string, error) {
-	ler, _, err := e.rl.Get()
+	ler, _, err := e.rl.Get(context.Background())
 	if err != nil {
 		return "", fmt.Errorf("failed to get leader election record: %v", err)
 	}